move Profile*Fragment to separate files
